Create people helper class
